move SimplePressureMetaTileEntity to api
