+++ src/main/java/gregicality/science/common/pipelike/pressure/BlockPressurePipe.java
@@ -48,12 +48,12 @@ public class BlockPressurePipe extends BlockMaterialPipe<PressurePipeType, Press
     }
 
     public static Cuboid6 getFlangeBox(EnumFacing side, double thickness) { // F**k you float
-        double min = (thickness > 0.3 ? (3 - 4 * thickness) / 8.0 : (7 - 8 * thickness) / 16.0) + 0.0002;
-        double max = 1.0 - min;
-        double faceMin = 0.0002;
-        double faceMax = 0.9998;
-        double flangeMin = 0.125;
-        double flangeMax = 0.875;
+        double min = (thickness > 0.3 ? (3.0d - 4.0d * thickness) / 8 : (7.0d - 8.0d * thickness) / 16) + 0.0002d;
+        double max = 1 - min;
+        double faceMin = 0.0002d;
+        double faceMax = 0.9998d;
+        double flangeMin = 0.125d;
+        double flangeMax = 0.875d;
 
         if (side == null) {
             return new Cuboid6(min, min, min, max, max, max);
